requirements.txt
alembic==1.16.5
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
Flask==3.1.2
flask-cors==6.0.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
MouseInfo==0.1.3
pillow==11.3.0
psycopg2-binary==2.9.11
# 图形界面自动化
pywinauto>=0.6.0
pyperclip>=1.8.0
pywin32>=300
requests==2.32.5
soupsieve==2.8
SQLAlchemy==2.0.45
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.6.1
Werkzeug==3.1.4
zipp==3.23.0
